The Great Refactoring
Move all of the page specific stuff out of main loop,
creating a single interface `Generate()` that handles
loading, parsing, converting, and output of a file.

This should probably be moved to a separate package with tests,
but for now we're relying on the `main` package test for integration.

One change slipped in here, so not a "pure" refactor.

Title of a Page inst has been moved to the Data struct,
the reason will be clear soon when first-class metadata is supported.

For now, all of the non-existent users can migrate their templates:

* {{ .Title }} is now {{ .Data.Title }}

Sorry for the breaking change :)