FrontMatter
This is first take at page meta-data at the top of a document,
you can add a chunk of YAML that will be parsed into the page's data.

First use for this is <title> in the HTML output.

For more ideas see how Jekyll does this:
https://jekyllrb.com/docs/frontmatter/

The regular expression used to parse this data was inspired by Jekyll:
https://github.com/jekyll/jekyll/blob/a944dd940a51db6b9a8db3f7f2dd15de8e04e764/lib/jekyll/document.rb#L13

Though this doesn't strip the data from the document,
will have to come back and fix this after.

Could also use some tests and maybe better error checking.

Lastly, this should also get moved to a package with unit tests.