Timestamps
{{ .Timestamp "my-date-string" }}

Really anything can go in there, no formatting of the given string.

So I'm using a timestamp fn in emacs to insert the string.

```lisp
(defun timestamp () (interactive)
       (insert (format-time-string "%Y-%m-%dT%H:%M:%S")))
```

Because if we generate a timestamp at build time, it would always change!

So the date should be static and included in the content of the post.

This could maybe be changed to a `Anchor` tag instead,
since that's all it really does.. create a named anchor in HTML.