Factored out assets into its own package and move config out of page
In order to remove the dependency on config from page, everything from
source, destination, template, and assets must be passed in before
generation.

This has an added benefit that fields aren't arbitrarily set during
load or known through proxy by member functions, as the case was for
source and destination.

Member functions that essentially decorate some data on an instance
were the give away. In this case, it's important to ask ourselves:

Will this data change through the lifecycle of the object?

If the answer is no, then it should be set on initialization.

Now we have less cruft in the page package and is easier to test.

Next, let's figure out how to refactor load and Generate(), then we to
do with the stuff sitting in main.