Fold source and destination into (*page.Page).Load()
While this adds more logic into load, it makes generate much simpler
and since that's where we'll be adding more stuff like generating an
index page and rss feed, etc. I want to keep it small.

Also, it seemed unnecessary to have a private method in one class,
that sets an attribute on an instance which is used inside a private
function within that namespace. Maybe those words don't quite explain
it well enough, but for sure it was confusing.

My goal from here is to make generation load files at once into a
slice that I can iterate over as many times as I wish for example
sorting by date and writing an archive page, or generating RSS.