Use the member function page.Title() for <title>
This function will try to determine the best possible option to call
the page by name.

This allows us to always print a <title>, whether dealing with a real
document or a proxy, like index. Or when the page specifies it's own
title in the meta-data.

Lastly, we probably want some way to define a site title in config and
append that to here.

Hmm.. how to do that without passing config around? Time to refactor
to use cobra/viper?! ...not so fast.