Move templates for index and page out into layouts
While this change means that users must implement the index template
themselves, it makes generation much simpler.

There is no migration path to move templates for this new behavior.

Thankfully no one but me is using this code!

The benefit is that I can now have separate templates for pages and
index is used for tags as well.