Use absolute paths when creating links
This resolves an issue where links to tags, of a page, from a tag
page... was causing a nesting issue.

Because it was using relative links until now.

If we want to keep using relative links, that would require creating
some kind of tree for pages to know how deeply nested they are, in
order to generate the proper links to assets and other places.

This solves the issue by allowing a page to always know the root.

It depends on whether config.SiteURL is filled out, and you probably
don't want to fill it out when testing local generation.

In that case use a blank string, or leave it empty, and all the links
will be fully resolved to wherever the output directory is.